# [Optimization] Add instructions for consuming non-newlines and advancing in scalar view