A simple new instruction to replace the consumers that are used for emitting `.any` and `.anyNonNewline` instructions in scalar and grapheme mode

Makes matching a default options `.` about ~3-5% faster
```
- ReluctantQuantWithTerminalWhole         7.8ms	8.2ms	-402µs		-4.9%
- ReluctantQuantWhole                     11.7ms	12ms	-345µs		-2.9%
```

Doesn't affect a lot of benchmarks in our suite because we mostly have quantified `.`s like `.*` and that ends up in the `.quantify` optimization
<img width="907" alt="image" src="https://user-images.githubusercontent.com/49693777/182730066-c0696be1-0c7e-4ad5-947d-3bcc4c03986f.png">