@milseman 2022-08-04
Is this substantially different than a consume builtin where anyNonNewline is a kind of builtin? I.e. is the dedicated opcode just a particular encoding of the operation?