@rctcwyvrn 2022-08-04
We could write it like this 
```  
mutating func emitAnyNonNewline() {
    let model = _CharacterClassModel.init(
      cc: .verticalWhitespace,
      options: options,
      isInverted: true)
    builder.buildMatchBuiltin(model: model)
  }
```

But since this falls into the switch statement with all the other builtin cases it ends up being much slower 

```
=== Regressions ======================================================================
- ReluctantQuantWhole                     14ms	12ms	2.01ms		16.7%
- ReluctantQuantWithTerminalWhole         10.1ms	8.2ms	1.88ms		22.9%
```