@milseman 2022-08-10
Right, so the choice of a different instruction is reorganizing the branching story. There are lots of ways to organize this, e.g. we could branch early for this case or we could even switch over the combination of op-code and builtin (and for non-builtin-taking operations, that would be zero-ed out).

Anyways, I'm fine taking this because it locks in the kind of performance story we will want at the end.